Shift all OCR content to a single module
